Refactor eslintignore, add backup script, update main.js
This commit refactors the eslintignore file to include all TypeScript
definition files in the /src/types directory, excludes the doc/ and
node_modules/ directories from linting, and adds the schema/ directory
to the list of ignored files. Additionally, a new backup script,
make-backup.mjs, is added to the backup/ directory. Finally, the main.js
file in the doc/assets/ directory undergoes minor code cleanup.

The changes improve linting accuracy and organization, introduce a
script for creating backups of the API app with data, and clean up the
main.js file.

No issue references.