feat: Update ESLint configuration and package versions
- Update the "source.fixAll.eslint" setting in the VSCode configuration
to use the value "explicit"
- Bump the version of the "3d-inventory-mongo-api" package in the
package-lock.json file from 0.0.12 to 0.0.13
- Update various devDependencies in the package-lock.json file,
including "@tsconfig/node21" to version 21.0.0 and
"@tsconfig/recommended" to version 1.0.3
- Upgrade "@typescript-eslint/eslint-plugin" to version 6.13.2 and
"@typescript-eslint/parser" to version 6.13.2

These changes improve the ESLint configuration and update the package
versions, ensuring compatibility and leveraging the latest features. No
significant consequences are expected. Issue #123

Note: Please replace "Issue #123" with the relevant issue number or
remove it if not applicable.