feat: Upgrade package version, create mongo API for 3d-inventory project (#123)
Upgrade the package versions in package-lock.json and package.json from
0.0.13 to 0.0.14. This upgrade includes various dependencies, such as
"@apollo/client" to "^3.8.8".

In addition, the README.md file has been updated to fix a typo and
provide a correct hyperlink for the "3d-inventory" project.

These changes are necessary to create a mongo API for the 3d-inventory
project, improving functionality and ensuring compatibility with the
latest dependencies.

Closes #123.