feat: Add configuration file and update environment variables
This commit adds a new configuration file for the project named
`.42c/conf.yaml`. Additionally, the environment variable `URI` in the
`.env.example` file has been replaced with `ATLAS_URI`, reflecting the
use of the MongoDB Atlas URI. These changes improve the organization and
clarity of the project configuration.

Fixes #123