feat: Add MongoDB connection functions and pre-defined data
This commit adds the necessary functions to connect to a MongoDB Atlas cluster and database. It also includes pre-defined data for attribute categories and components. The new database queries for inserting devices and models have been moved to a "queries" directory for better organization. This change enables seamless interaction with the MongoDB database and provides a foundation for populating data in the application.

Issue #123